Remove example Jupyter notebook and associated image
Deleted `notebooks/example.ipynb` and `notebooks/boxplot.png` to clean up outdated or unnecessary files from the repository. This will help maintain a more organized and relevant codebase.
