+++ .idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.11 (code)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="obia" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12 (obia) (2)" project-jdk-type="Python SDK" />
 </project>
