+++ README.md
@@ -1,2 +1,3 @@
-# obia
-Object Based Image Analysis
+# obia: Object Based Image Analysis
+
+This is still an experimental library.
