+++ requirements.txt
@@ -9,3 +9,5 @@ scikit-image~=0.23.1
 tqdm~=4.66.2
 scikit-learn~=1.4.2
 affine~=2.4.0
+pyproj~=3.6.1
+shap~=0.46.0
